Moving connection manager to include
Because the user will need to extend connection manager and instantiate connection managers, this should not be hidden in private.
